create post

+++ app/Http/Controllers/AdminUserController.php
@@ -124,7 +124,9 @@ class AdminUserController extends Controller
     public function destroy($id)
     {
         $user = User::findorfail($id);
-        unlink(public_path().$user->photo->file);
+        if ($user->photo) {
+            unlink(public_path() . $user->photo->file);
+        }
         $user->delete();
 
         Session::flash('deleted_user','The user had been deleted.');
